backend/requirements.txt
Flask==2.2.3
Flask-Cors==4.0.0
azure-core==1.30.1
azure-search-documents==11.4.0
openai==1.14.3
spacy==3.7.4
gunicorn==21.2.0
Werkzeug==2.2.3
